Add support for password autofill during login
Fixes #7615

Password autofill may enter usernames instead of email, as there is no concrete way to tell iOS to prefer email over usernames when both are available. To deal with that possibility, an invisible password field has been added that can only be filled by password autofill. When it gets a value, login is attempted immediately.