Move all files in `OAuth/` to `GoogleSignIn/`
I realized a while ago that trying to clearly separate the OAuth
implementation from the Google SignIn code was a distraction from the
goal of getting this project done ASAP and an unnecessary constraint.

We can always spend some time defining a clearer separation in the
future, when we'll add other authentication methods that support OAuth.