Remove method forwarding the call to handle an URL to the Google SDK
The question was whether we should implement some kind of handling on
our hand, but that seems unnecessary given Jetpack and WordPress haven't
shown any issues without it.

Besides, what could we return if not `false`? We no longer have the
Google SDK available, so we cannot "handle" the URL in that way.

Granted, there could be other ways to handle it consistent with our
library implementation, but they don't seem necessary at this time.