POC to show that setting the currency overrides user default
On multicurrency sites, the user’s default currency is used for order creation when none is specified in the API request.

This may differ from the site’s default currency. A user can set their default currency at `/my-account/edit-account/`.

If their default is different from the site’s, they won’t be able to use in-person payments.

Additionally, order creation will show that they’re creating the order in the site’s default currency, but when it’s created it’ll actually use the user’s, so it’ll appear to switch over which is confusing.

To work around this in a world where the app doesn’t really support multicurrency, we can set the site’s default currency code in the initial order creation request.