Debug and prevent error on pull to refresh
The reason it was showing and error is that we set the state to `.loading` when you pull to refresh, and that redrew the view with the refreshable control: ItemListView.

It redrew `ItemListView` because `ItemListView` used the `posModel.itemsViewState.itemsStack.root` in its body, to pass it to the `ItemList` (among other things.)

PTR set `itemsStack.root = .loading` at the start of its refresh

SwiftUI can’t tell that the `ItemListView`’s dependency on the `itemsStack.root` is only to pass it to a subview, so it registers `ItemListView` to be redrawn whenever the root stack changes.

Removing those uses from `ItemListView`, and having `ItemList` declare its own dependency on the root stack via the `posModel` environment object, prevents the ItemListView (and importantly, its refreshable,) from being redrawn for a root stack update, only the list content is redrawn.

Once that’s not redrawn, the error no longer shows.

This is all academic, because we don’t need to set the root state to `loading` for PTR, we should just leave it alone and let the SwiftUI provided spinner show that it’s loading. Setting the state to loading adds the ghost cell at the end of the list, and that isn’t really what we want for a PTR with items.