Fold NSURL+IDN into the project — Remove CP dependency
For the record, I tried using `git filter-repo` but the result it gave
me was a single commit.

The command I used from the NSURL+IDN repo was:

```
git filter-repo \
  --path Sources \
  --path-rename Sources:WordPressAuthenticator/Sources \
  --force
```

The result:

```
$ git log

commit 5e4a19a43dad69e1923b51c3697ed7791e4030e2 (HEAD -> trunk)
Author: Tony Li <tony.li@automattic.com>
Date:   Thu Nov 24 10:10:31 2022 +1300

    Add Swift Package Manager support
```

No other commit in the logs.

From the repo's main branch, `git log --follow -- Sources` produces the
same result.

I decided not to dig more into it.